requirements.txt
flask==2.3.3
flask_oauthlib
beautifulsoup4==4.13.3
requests==2.32.3
selenium==4.29.0
mysql-connector-python==9.2.0
python-dotenv==1.0.1
webdriver-manager==4.0.1
lxml==5.1.0
urllib3==2.3.0
bs4==0.0.2
soupsieve==2.6
typing_extensions==4.12.2
charset-normalizer==3.4.1
h11==0.14.0
attrs==25.1.0
PySocks==1.7.1
filelock==3.17.0
cffi==1.17.1
trio==0.29.0
outcome==1.3.0.post0
sniffio==1.3.1
sortedcontainers==2.4.0
idna==3.10
MarkupSafe==3.0.2
werkzeug==2.3.7
click==8.1.8
Jinja2==3.1.5
blinker==1.9.0
itsdangerous==2.2.0